requirements.txt
Flask
Flask-Cors
requests
beautifulsoup4
openai
python-dotenv